Refactor code to follow SRP
